+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Changelog
 =========
+
+## Unreleased
+### Added
+- `flipbox\patron\services\ManageTokens::find()` will find a token based on another AccessToken
+
 ## 1.0.0-rc.1 - 2018-04-04
 ### Fixed
 - Bugs when attempting to get a token from an `AbstractProvider`